Added deep sort torch handler

+++ detector/prediction_status.h
@@ -9,7 +9,7 @@ struct PredictionStatus {
         : found(false) {
     }
     bool found;
-    std::vector<cv::Rect2d> rects;
+    std::vector<cv::Rect2f> rects;
     cv::Mat frame;
 };
